Split endpointing into per-pipeline classes via composition
