+++ tests/collections/asr/inference/test_greedy_endpointing.py
@@ -15,8 +15,14 @@
 import pytest
 import torch
 
-from nemo.collections.asr.inference.streaming.endpointing.greedy.greedy_ctc_endpointing import CTCGreedyEndpointing
-from nemo.collections.asr.inference.streaming.endpointing.greedy.greedy_rnnt_endpointing import RNNTGreedyEndpointing
+from nemo.collections.asr.inference.streaming.endpointing.greedy.buffered_ctc_endpointer import BufferedCTCEndpointer
+from nemo.collections.asr.inference.streaming.endpointing.greedy.buffered_rnnt_endpointer import BufferedRNNTEndpointer
+from nemo.collections.asr.inference.streaming.endpointing.greedy.cache_aware_ctc_endpointer import (
+    CacheAwareCTCEndpointer,
+)
+from nemo.collections.asr.inference.streaming.endpointing.greedy.cache_aware_rnnt_endpointer import (
+    CacheAwareRNNTEndpointer,
+)
 from nemo.collections.asr.inference.utils.endpointing_utils import millisecond_to_frames
 
 # Vocabulary used by the detect_eou tests below.
@@ -25,7 +31,8 @@ from nemo.collections.asr.inference.utils.endpointing_utils import millisecond_t
 EOU_VOCAB = ["▁hello", "world", "▁the", ".", ","]
 EOU_BLANK = len(EOU_VOCAB)
 EOU_ABSORB_IDS = {3, 4}
-ENDPOINTING_CLASSES = [CTCGreedyEndpointing, RNNTGreedyEndpointing]
+# detect_eou_in_buffer is the cache-aware algorithm, shared by the CTC and RNNT cache-aware endpointers.
+ENDPOINTING_CLASSES = [CacheAwareCTCEndpointer, CacheAwareRNNTEndpointer]
 
 
 def _make_detect_eou_endpointer(endpointing_cls, ms_per_timestep=20, stop_history_eou=80, stop_history_eou_end=None):
@@ -40,7 +47,7 @@ def _make_detect_eou_endpointer(endpointing_cls, ms_per_timestep=20, stop_histor
 
 
 class TestDetectEou:
-    """Tests for the full-buffer EoU detection (GreedyEndpointing.detect_eou)."""
+    """Tests for the cache-aware EoU detection (CacheAwareEndpointer.detect_eou_in_buffer)."""
 
     @pytest.mark.unit
     def test_trailing_silence(self):
@@ -342,132 +349,111 @@ class TestGreedyEndpointing:
 
     @pytest.mark.unit
     def test_endpointing_with_negative_stop_history_eou(self):
-        for endpointing_cls in [CTCGreedyEndpointing, RNNTGreedyEndpointing]:
-            greedy_endpointing = endpointing_cls(vocabulary=["a", "b", "c"], ms_per_timestep=100, stop_history_eou=-1)
-            if isinstance(greedy_endpointing, CTCGreedyEndpointing):
-                b = len(greedy_endpointing.greedy_ctc_decoder.vocabulary)
-            else:
-                b = len(greedy_endpointing.greedy_rnnt_decoder.vocabulary)
-            emissions = [0, 1, 2, b, b, b, b, b, b, b, b, b]
+        # detect_eou_given_emissions is the buffered-CTC algorithm.
+        ep = BufferedCTCEndpointer(vocabulary=["a", "b", "c"], ms_per_timestep=100, stop_history_eou=-1)
+        b = ep.token_classifier.blank_id
+        emissions = [0, 1, 2, b, b, b, b, b, b, b, b, b]
 
-            # False case, because stop_history_eou = -1
-            assert greedy_endpointing.detect_eou_given_emissions(emissions, 3) == (False, -1)
+        # False case, because stop_history_eou = -1
+        assert ep.detect_eou_given_emissions(emissions, 3) == (False, -1)
 
     @pytest.mark.unit
     def test_endpointing_with_positive_stop_history_eou(self):
-        for endpointing_cls in [CTCGreedyEndpointing, RNNTGreedyEndpointing]:
-            greedy_endpointing = endpointing_cls(
-                vocabulary=["a", "b", "c"], ms_per_timestep=20, stop_history_eou=100, residue_tokens_at_end=0
-            )
-            if isinstance(greedy_endpointing, CTCGreedyEndpointing):
-                b = len(greedy_endpointing.greedy_ctc_decoder.vocabulary)
-            else:
-                b = len(greedy_endpointing.greedy_rnnt_decoder.vocabulary)
-            emissions = [0, 1, 2, b, b, b, b, b, b, b, b, b]
+        ep = BufferedCTCEndpointer(
+            vocabulary=["a", "b", "c"], ms_per_timestep=20, stop_history_eou=100, residue_tokens_at_end=0
+        )
+        b = ep.token_classifier.blank_id
+        emissions = [0, 1, 2, b, b, b, b, b, b, b, b, b]
 
-            for pivot_point in range(len(emissions)):
-                eou_detected, eou_detected_at = greedy_endpointing.detect_eou_given_emissions(emissions, pivot_point)
-                assert eou_detected == True
+        for pivot_point in range(len(emissions)):
+            eou_detected, eou_detected_at = ep.detect_eou_given_emissions(emissions, pivot_point)
+            assert eou_detected == True
 
     @pytest.mark.unit
     def test_detect_eou_given_timestamps_empty_inputs(self):
-        for endpointing_cls in [CTCGreedyEndpointing, RNNTGreedyEndpointing]:
-            greedy_endpointing = endpointing_cls(
-                vocabulary=["a", "b", "c"], ms_per_timestep=80, stop_history_eou=100, residue_tokens_at_end=0
-            )
+        # detect_eou_given_timestamps is the buffered-RNNT algorithm.
+        ep = BufferedRNNTEndpointer(
+            vocabulary=["a", "b", "c"], ms_per_timestep=80, stop_history_eou=100, residue_tokens_at_end=0
+        )
 
-            # Test with empty timesteps and tokens
-            timesteps = torch.tensor([])
-            tokens = torch.tensor([])
-            alignment_length = 10
+        # Test with empty timesteps and tokens
+        timesteps = torch.tensor([])
+        tokens = torch.tensor([])
+        alignment_length = 10
 
-            eou_detected, eou_detected_at = greedy_endpointing.detect_eou_given_timestamps(
-                timesteps, tokens, alignment_length
-            )
-            assert eou_detected == False
-            assert eou_detected_at == -1
+        eou_detected, eou_detected_at = ep.detect_eou_given_timestamps(timesteps, tokens, alignment_length)
+        assert eou_detected == False
+        assert eou_detected_at == -1
 
     @pytest.mark.unit
     def test_detect_eou_given_timestamps_disabled_stop_history(self):
-        for endpointing_cls in [CTCGreedyEndpointing, RNNTGreedyEndpointing]:
-            greedy_endpointing = endpointing_cls(
-                vocabulary=["a", "b", "c"],
-                ms_per_timestep=80,
-                stop_history_eou=-1,  # Disabled
-                residue_tokens_at_end=0,
-            )
+        ep = BufferedRNNTEndpointer(
+            vocabulary=["a", "b", "c"],
+            ms_per_timestep=80,
+            stop_history_eou=-1,  # Disabled
+            residue_tokens_at_end=0,
+        )
 
-            timesteps = torch.tensor([0, 2, 4, 6])
-            tokens = torch.tensor([0, 1, 2, 3])
-            alignment_length = 10
+        timesteps = torch.tensor([0, 2, 4, 6])
+        tokens = torch.tensor([0, 1, 2, 3])
+        alignment_length = 10
 
-            eou_detected, eou_detected_at = greedy_endpointing.detect_eou_given_timestamps(
-                timesteps, tokens, alignment_length
-            )
-            assert eou_detected == False
-            assert eou_detected_at == -1
+        eou_detected, eou_detected_at = ep.detect_eou_given_timestamps(timesteps, tokens, alignment_length)
+        assert eou_detected == False
+        assert eou_detected_at == -1
 
     @pytest.mark.unit
     def test_detect_eou_given_timestamps_trailing_silence(self):
-        for endpointing_cls in [CTCGreedyEndpointing, RNNTGreedyEndpointing]:
-            greedy_endpointing = endpointing_cls(
-                vocabulary=["a", "b", "c"], ms_per_timestep=20, stop_history_eou=80, residue_tokens_at_end=0
-            )
+        ep = BufferedRNNTEndpointer(
+            vocabulary=["a", "b", "c"], ms_per_timestep=20, stop_history_eou=80, residue_tokens_at_end=0
+        )
 
-            # Last token at position 5, alignment_length is 10
-            # Trailing silence = 10 - 4 - 1 = 5 frames > stop_history_eou (4)
-            timesteps = torch.tensor([0, 1, 2, 3, 4])
-            tokens = torch.tensor([0, 1, 2, 3, 4])
-            alignment_length = 10
+        # Last token at position 5, alignment_length is 10
+        # Trailing silence = 10 - 4 - 1 = 5 frames > stop_history_eou (4)
+        timesteps = torch.tensor([0, 1, 2, 3, 4])
+        tokens = torch.tensor([0, 1, 2, 3, 4])
+        alignment_length = 10
 
-            eou_detected, eou_detected_at = greedy_endpointing.detect_eou_given_timestamps(
-                timesteps, tokens, alignment_length
-            )
-            assert eou_detected == True
-            # eou_detected_at = 4 + 1 + 4//2 = 7
-            assert eou_detected_at == 7
+        eou_detected, eou_detected_at = ep.detect_eou_given_timestamps(timesteps, tokens, alignment_length)
+        assert eou_detected == True
+        # eou_detected_at = 4 + 1 + 4//2 = 7
+        assert eou_detected_at == 7
 
     @pytest.mark.unit
     def test_detect_eou_given_timestamps_no_trailing_silence(self):
-        for endpointing_cls in [CTCGreedyEndpointing, RNNTGreedyEndpointing]:
-            greedy_endpointing = endpointing_cls(
-                vocabulary=["a", "b", "c"], ms_per_timestep=20, stop_history_eou=80, residue_tokens_at_end=0
-            )
+        ep = BufferedRNNTEndpointer(
+            vocabulary=["a", "b", "c"], ms_per_timestep=20, stop_history_eou=80, residue_tokens_at_end=0
+        )
 
-            # Last token at position 8, alignment_length is 10
-            # Trailing silence = 10 - 8 - 1 = 1 frame < stop_history_eou (4)
-            timesteps = torch.tensor([0, 1, 2, 3, 8])
-            tokens = torch.tensor([0, 1, 2, 3, 4])
-            alignment_length = 10
+        # Last token at position 8, alignment_length is 10
+        # Trailing silence = 10 - 8 - 1 = 1 frame < stop_history_eou (4)
+        timesteps = torch.tensor([0, 1, 2, 3, 8])
+        tokens = torch.tensor([0, 1, 2, 3, 4])
+        alignment_length = 10
 
-            eou_detected, eou_detected_at = greedy_endpointing.detect_eou_given_timestamps(
-                timesteps, tokens, alignment_length
-            )
-            assert eou_detected == False
-            assert eou_detected_at == -1
+        eou_detected, eou_detected_at = ep.detect_eou_given_timestamps(timesteps, tokens, alignment_length)
+        assert eou_detected == False
+        assert eou_detected_at == -1
 
     @pytest.mark.unit
     def test_detect_eou_given_timestamps_gap_detection(self):
-        for endpointing_cls in [CTCGreedyEndpointing, RNNTGreedyEndpointing]:
-            greedy_endpointing = endpointing_cls(
-                vocabulary=["a", "b", "c"], ms_per_timestep=20, stop_history_eou=80, residue_tokens_at_end=0
-            )
+        ep = BufferedRNNTEndpointer(
+            vocabulary=["a", "b", "c"], ms_per_timestep=20, stop_history_eou=80, residue_tokens_at_end=0
+        )
 
-            # Large gap between tokens: 8 - 2 - 1 = 5 frames > stop_history_eou (4)
-            timesteps = torch.tensor([0, 2, 8, 9])
-            tokens = torch.tensor([0, 1, 2, 3])
-            alignment_length = 10
+        # Large gap between tokens: 8 - 2 - 1 = 5 frames > stop_history_eou (4)
+        timesteps = torch.tensor([0, 2, 8, 9])
+        tokens = torch.tensor([0, 1, 2, 3])
+        alignment_length = 10
 
-            eou_detected, eou_detected_at = greedy_endpointing.detect_eou_given_timestamps(
-                timesteps, tokens, alignment_length
-            )
-            assert eou_detected == True
-            # eou_detected_at = 2 + 1 + 4//2 = 5
-            assert eou_detected_at == 5
+        eou_detected, eou_detected_at = ep.detect_eou_given_timestamps(timesteps, tokens, alignment_length)
+        assert eou_detected == True
+        # eou_detected_at = 2 + 1 + 4//2 = 5
+        assert eou_detected_at == 5
 
     @pytest.mark.unit
     def test_rnnt_vad_endpointing_disabled(self):
-        rnnt_endpointing = RNNTGreedyEndpointing(
+        rnnt_endpointing = BufferedRNNTEndpointer(
             vocabulary=["a", "b", "c"],
             ms_per_timestep=100,
             effective_buffer_size_in_secs=None,  # VAD disabled
@@ -484,7 +470,7 @@ class TestGreedyEndpointing:
 
     @pytest.mark.unit
     def test_rnnt_vad_endpointing_enabled_no_eou(self):
-        rnnt_endpointing = RNNTGreedyEndpointing(
+        rnnt_endpointing = BufferedRNNTEndpointer(
             vocabulary=["a", "b", "c"],
             ms_per_timestep=100,
             effective_buffer_size_in_secs=2.0,  # VAD enabled
@@ -500,7 +486,7 @@ class TestGreedyEndpointing:
 
     @pytest.mark.unit
     def test_rnnt_vad_endpointing_enabled_with_eou(self):
-        rnnt_endpointing = RNNTGreedyEndpointing(
+        rnnt_endpointing = BufferedRNNTEndpointer(
             vocabulary=["a", "b", "c"],
             ms_per_timestep=100,
             effective_buffer_size_in_secs=2.0,  # VAD enabled
@@ -519,7 +505,7 @@ class TestGreedyEndpointing:
 
     @pytest.mark.unit
     def test_rnnt_vad_endpointing_enabled_with_eou_at_end(self):
-        rnnt_endpointing = RNNTGreedyEndpointing(
+        rnnt_endpointing = BufferedRNNTEndpointer(
             vocabulary=["a", "b", "c"],
             ms_per_timestep=100,
             effective_buffer_size_in_secs=2.0,  # VAD enabled
